refractor empty pool cases into its module
